Stage 2 Project
I send 3 files. For the python html generator, I used the raw version
of my notes to create a list, and then I adapted the html generator
coded in class to my html format and I developed it to take not only
one title and concept, but the whole notes of the session. Then I
refined the HTML and CSS files (so the wording of the notes might be
slightly different).

P.S.: the project instruction for the phyton file were not super clear,
I hope this is fine